requirements.txt
alabaster==0.7.12
alembic==0.9.9
altgraph==0.16.1
Babel==2.7.0
blinker==1.4
cachetools==4.1.1
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.4.1
cryptography==2.9.2
docutils==0.14
Flask==1.0.2
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.11.3
idna==2.6
imagesize==1.1.0
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
macholib==1.11
Mako==1.0.7
MarkupSafe==1.0
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==2.0.7
packaging==19.0
pefile==2019.4.18
Pillow==6.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.4.2
PyInstaller==3.6
pyOpenSSL==19.1.0
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
python-dateutil==2.7.2
python-editor==1.0.3
python-google-calendar-api==0.5
pytz==2019.1
pywin32-ctypes==0.2.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==4.6
six==1.11.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
uritemplate==3.0.1
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
xlrd==1.2.0
